requirements.txt
# Core dependencies
numpy==1.26.4
protobuf==3.20.3
pillow==10.4.0
opencv-python==4.10.0.84

# Torch (CUDA 12.1 stack)
--index-url https://download.pytorch.org/whl/cu121
torch==2.4.0+cu121
torchvision==0.19.0+cu121
torchaudio==2.4.0+cu121

# Diffusion & ML stack
diffusers==0.30.0
transformers==4.43.3
accelerate==0.33.0
huggingface_hub==0.24.6
safetensors==0.4.3
timm==1.0.19
einops==0.8.1

# Image generation helpers
controlnet_aux==0.0.7
rembg==2.0.56

# CLI & utilities
typer==0.12.3
click==8.1.7
rich==13.9.4
PyYAML==6.0.2
six==1.17.0
